Code review changes:
* Create a JsonObjectMarshaller interface instead of using LineAggregator
* Introduce a base class named AbstractFileItemWriter for file writers
(with two sub-classes: FlatFileItemWriter and (the new) JsonFileItemWriter)
