+++ spring-batch-infrastructure/src/main/java/org/springframework/batch/item/json/JsonFileItemWriter.java
@@ -22,15 +22,16 @@ import java.util.List;
 
 import org.springframework.batch.item.WriteFailedException;
 import org.springframework.batch.item.WriterNotOpenException;
-import org.springframework.batch.item.file.FlatFileItemWriter;
-import org.springframework.batch.item.file.transform.LineAggregator;
+import org.springframework.batch.item.support.AbstractFileItemWriter;
 import org.springframework.core.io.Resource;
 import org.springframework.util.Assert;
+import org.springframework.util.ClassUtils;
 
 /**
- * Item writer that writes data in json format to an output file.
- * Items are transformed to json format using a {@link LineAggregator}.
- * Items will be enclosed in a json array as follows:
+ * Item writer that writes data in json format to an output file. The location
+ * of the output file is defined by a {@link Resource} and must represent a
+ * writable file. Items are transformed to json format using a
+ * {@link JsonObjectMarshaller}. Items will be enclosed in a json array as follows:
  *
  * <p>
  * <code>
@@ -42,30 +43,55 @@ import org.springframework.util.Assert;
  * </code>
  * </p>
  *
- * @see GsonLineAggregator
- * @see JacksonLineAggregator
+ * The implementation is <b>not</b> thread-safe.
+ *
+ * @see GsonJsonObjectMarshaller
+ * @see JacksonJsonObjectMarshaller
  * @param <T> type of object to write as json representation
  * @author Mahmoud Ben Hassine
  * @since 4.1
  */
-public class JsonItemWriter<T> extends FlatFileItemWriter<T> {
+public class JsonFileItemWriter<T> extends AbstractFileItemWriter<T> {
 
 	private static final char JSON_OBJECT_SEPARATOR = ',';
 	private static final char JSON_ARRAY_START = '[';
 	private static final char JSON_ARRAY_STOP = ']';
 
+	private JsonObjectMarshaller<T> jsonObjectMarshaller;
+
 	/**
-	 * Create a new {@link JsonItemWriter} instance.
+	 * Create a new {@link JsonFileItemWriter} instance.
 	 * @param resource to write json data to
-	 * @param lineAggregator used to aggregate object into json representation
+	 * @param jsonObjectMarshaller used to marshal object into json representation
 	 */
-	public JsonItemWriter(Resource resource, LineAggregator<T> lineAggregator) {
+	public JsonFileItemWriter(Resource resource, JsonObjectMarshaller<T> jsonObjectMarshaller) {
 		Assert.notNull(resource, "resource must not be null");
-		Assert.notNull(lineAggregator, "line aggregator must not be null");
+		Assert.notNull(jsonObjectMarshaller, "json object marshaller must not be null");
 		setResource(resource);
-		setLineAggregator(lineAggregator);
+		setJsonObjectMarshaller(jsonObjectMarshaller);
 		setHeaderCallback(writer -> writer.write(JSON_ARRAY_START));
 		setFooterCallback(writer -> writer.write(this.lineSeparator + JSON_ARRAY_STOP + this.lineSeparator));
+		setExecutionContextName(ClassUtils.getShortName(JsonFileItemWriter.class));
+	}
+
+	/**
+	 * Assert that mandatory properties (jsonObjectMarshaller) are set.
+	 *
+	 * @see org.springframework.beans.factory.InitializingBean#afterPropertiesSet()
+	 */
+	@Override
+	public void afterPropertiesSet() throws Exception {
+		if (this.append) {
+			this.shouldDeleteIfExists = false;
+		}
+	}
+
+	/**
+	 * Set the {@link JsonObjectMarshaller} to use to marshal object to json.
+	 * @param jsonObjectMarshaller the marshaller to use
+	 */
+	public void setJsonObjectMarshaller(JsonObjectMarshaller<T> jsonObjectMarshaller) {
+		this.jsonObjectMarshaller = jsonObjectMarshaller;
 	}
 
 	@Override
@@ -88,7 +114,7 @@ public class JsonItemWriter<T> extends FlatFileItemWriter<T> {
 				lines.append(JSON_OBJECT_SEPARATOR).append(this.lineSeparator);
 			}
 			T item = iterator.next();
-			lines.append(' ').append(this.lineAggregator.aggregate(item));
+			lines.append(' ').append(this.jsonObjectMarshaller.marshal(item));
 			if (iterator.hasNext()) {
 				lines.append(JSON_OBJECT_SEPARATOR).append(this.lineSeparator);
 			}
